Updated user api endpoint and response
