+++ backend/routes/volunteerHubRoutes.js
@@ -1,6 +1,14 @@
 const axios = require("axios");
 const keys = require("../config/keys");
 const stub = require("../services/stubs.js");
+const { processUser } = require("../services/processRequests.js");
+
+const requestOptions = {
+  auth: {
+    username: keys.volunteerHubUsername,
+    password: keys.volunteerHubPassword
+  }
+};
 
 module.exports = app => {
   //GET EVENTS LIST FROM VOLUNTEER HUB
@@ -12,33 +20,23 @@ module.exports = app => {
       startDate +
       latestTime;
 
-    console.log("Get Request : " + url);
+    axios.get(url, requestOptions).then(
+      request => {
+        var { data } = request;
+        data.map(event => {
+          const eventUID = event.EventUid;
 
-    axios
-      .get(url, {
-        auth: {
-          username: keys.volunteerHubUsername,
-          password: keys.volunteerHubPassword
-        }
-      })
-      .then(
-        request => {
-          //console.log("=== NEW EVENT REQUEST ===");
-          var { data } = request;
-          data.map(event => {
-            const eventUID = event.EventUid;
+          event.UserGroupRegistrations.map(async usergroup => {
+            const groupUID = usergroup.UserGroupUid;
 
-            event.UserGroupRegistrations.map(async usergroup => {
-              const groupUID = usergroup.UserGroupUid;
-
-              const userGroupStub = await stub.getStub(eventUID, groupUID);
-              //console.log(userGroupStub.url_stub);
-            });
+            const userGroupStub = await stub.getStub(eventUID, groupUID);
+            //console.log(userGroupStub.url_stub);
           });
-          res.send(data);
-        },
-        error => console.log(error)
-      );
+        });
+        res.send(data);
+      },
+      error => console.log(error)
+    );
   });
 
   //GET EVENTS BY ID FROM VOLUNTEER HUB
@@ -46,15 +44,8 @@ module.exports = app => {
     const eventID = req.query.eventID;
     url = "https://denver.volunteerhub.com/api/v1/events/" + eventID;
 
-    //console.log("Get Request : " + url);
-
     axios
-      .get(url, {
-        auth: {
-          username: keys.volunteerHubUsername,
-          password: keys.volunteerHubPassword
-        }
-      })
+      .get(url, requestOptions)
       .then(request => res.send(request.data), error => console.log(error));
   });
 
@@ -63,16 +54,12 @@ module.exports = app => {
     const id = req.query.id;
     url = "https://denver.volunteerhub.com/api/v2/users/" + id;
 
-    //console.log("Get Request : " + url);
-
-    axios
-      .get(url, {
-        auth: {
-          username: keys.volunteerHubUsername,
-          password: keys.volunteerHubPassword
-        }
-      })
-      .then(request => res.send(request.data), error => console.log(error));
+    axios.get(url, requestOptions).then(
+      request => {
+        res.send(processUser(request.data));
+      },
+      error => console.log(error)
+    );
   });
 
   //GET USERGROUPS BY PAGE FROM VOLUNTEER HUB
@@ -85,25 +72,11 @@ module.exports = app => {
 
     //console.log("Get Request : " + url);
 
-    axios
-      .get(url, {
-        auth: {
-          username: keys.volunteerHubUsername,
-          password: keys.volunteerHubPassword
-        }
-      })
-      .then(
-        request => res.send(request.data),
-        error => {
-          console.log(
-            "Error Code: " +
-              error.response.status +
-              " " +
-              error.response.statusText +
-              " - " +
-              url
-          );
-        }
-      );
+    axios.get(url, requestOptions).then(
+      request => res.send(request.data),
+      error => {
+        console.log(error);
+      }
+    );
   });
 };
