+++ client/src/containers/UserDetail.js
@@ -1,25 +1,22 @@
 import React from "react";
 import Grid from "@material-ui/core/Grid";
 import Typography from "@material-ui/core/Typography";
-import { userDataExtract } from "../utils/index";
 
 const UserDetail = props => {
   const { user } = props;
-
-  const userObject = userDataExtract(user);
 
   return (
     <Grid container>
       <Grid item xs={10} sm={3}>
         <Typography>
-          {userObject.FirstName} {userObject.LastName}
+          {user.FirstName} {user.LastName}
         </Typography>
       </Grid>
       <Grid item xs={12} sm={3}>
-        <Typography>{userObject.Phone} </Typography>
+        <Typography>{user.Phone} </Typography>
       </Grid>
       <Grid item xs={12} sm={3}>
-        <Typography>{userObject.Email} </Typography>
+        <Typography>{user.Email} </Typography>
       </Grid>
     </Grid>
   );
