+++ client/src/reducers/reducerUser.js
@@ -1,16 +1,13 @@
-import _ from "lodash";
 import { FETCH_USER, CLEAR_USERS } from "../actions/fetchUser";
 
 export default function(state = {}, action) {
   switch (action.type) {
     case FETCH_USER:
       if (action.payload.data) {
+        console.log(action.payload.data);
         return {
           ...state,
-          [action.payload.data.UserUid]: _.mapKeys(
-            action.payload.data.FormAnswers,
-            "FormQuestionUid"
-          )
+          [action.payload.data.UserUid]: action.payload.data
         };
       } else {
         return state;
